requirements.txt
flask==3.1.0
python-dotenv==1.0.1
sentence-transformers==3.3.1
PyMuPDF==1.25.3
google-generativeai==0.8.3
numpy==1.26.4
scikit-learn==1.6.1
gunicorn==23.0.0